+++ applications/fishing-map/src/features/sidebar/CategoryTabs.tsx
@@ -18,6 +18,7 @@ import useViewport from 'features/map/map-viewport.hooks'
 import FeedbackModal from 'features/feedback/FeedbackModal'
 import LanguageToggle from 'features/i18n/LanguageToggle'
 import LocalStorageLoginLink from 'routes/LoginLink'
+import HintsHub from 'features/help/hints/HintsHub'
 import styles from './CategoryTabs.module.css'
 
 const DEFAULT_WORKSPACE_LIST_VIEWPORT = {
@@ -55,9 +56,6 @@ function CategoryTabs({ onMenuClick }: CategoryTabsProps) {
   // const [modalHelpOpen, setModalHelpOpen] = useState(false)
   const [modalFeedbackOpen, setModalFeedbackOpen] = useState(false)
 
-  // const onHelpClick = () => {
-  //   setModalHelpOpen(true)
-  // }
   const onFeedbackClick = () => {
     setModalFeedbackOpen(true)
   }
@@ -95,11 +93,9 @@ function CategoryTabs({ onMenuClick }: CategoryTabsProps) {
           </li>
         ))}
         <div className={styles.separator}></div>
-        {/* <li className={cx(styles.tab, styles.secondary)}>
-          <button className={styles.tabContent} onClick={onHelpClick}>
-            <Icon icon="help" />
-          </button>
-        </li> */}
+        <li className={cx(styles.tab, styles.secondary)}>
+          <HintsHub />
+        </li>
         {userData && (
           <li className={cx(styles.tab, styles.secondary)}>
             <IconButton