+++ applications/fishing-map/src/features/timebar/Timebar.tsx
@@ -25,6 +25,7 @@ import { selectActivityCategory, selectTimebarGraph } from 'features/app/app.sel
 import { getEventLabel } from 'utils/analytics'
 import { upperFirst } from 'utils/info'
 import { selectIsMapDrawing } from 'features/map/map.selectors'
+import Hint from 'features/help/hints/Hint'
 import {
   setHighlightedTime,
   disableHighlightedTime,
@@ -200,7 +201,7 @@ const TimebarWrapper = () => {
   if (!start || !end || isMapDrawing) return null
 
   return (
-    <div>
+    <div className={styles.timebarWrapper}>
       <TimebarComponent
         enablePlayback={true}
         labels={labels}
@@ -261,6 +262,7 @@ const TimebarWrapper = () => {
         ) : null}
       </TimebarComponent>
       {!isSmallScreen && <TimebarSettings />}
+      <Hint id="changingTheTimeRange" className={styles.helpHint} />
     </div>
   )
 }