+++ applications/vessel-history/src/features/profile/components/activity/ActivityModalContentDetailsEncounter.tsx
@@ -1,8 +1,13 @@
-import React, { Fragment } from 'react'
+import React, { Fragment, useCallback, useState } from 'react'
 import { useTranslation } from 'react-i18next'
+import { useDispatch } from 'react-redux'
+import { redirect } from 'redux-first-router'
 import { EventVessel } from '@globalfishingwatch/api-types/dist'
+import { Spinner } from '@globalfishingwatch/ui-components'
 import { DEFAULT_EMPTY_VALUE } from 'data/config'
 import { RenderedEvent } from 'features/vessels/activity/vessels-activity.selectors'
+import { PROFILE } from 'routes/routes'
+import { useSearchConnect } from 'features/search/search.hooks'
 import ActivityModalContentField from './ActivityModalContentField'
 import ActivityModalContentDetails from './ActivityModalContentDetails'
 import styles from './ActivityModalDetails.module.css'
@@ -17,6 +22,46 @@ const ActivityModalContentDetailsEncounter: React.FC<ActivityModalContentProps> 
   const event = props.event
   const relatedVessel = event.encounter?.vessel as EventVessel
   const { t } = useTranslation()
+  const { findVessel } = useSearchConnect()
+  const dispatch = useDispatch()
+  const [profileLoading, setProfileLoading] = useState(false)
+
+  const openVesselProfile = useCallback(
+    async (vessel) => {
+      setProfileLoading(true)
+      let dataset = 'public-global-fishing-vessels:v20201001';
+      let vesselMatchId = null;
+      const vesselFound = await findVessel(vessel.id, vessel.name, vessel.flag, vessel.ssvid)
+      if (vesselFound) {
+        dataset = vesselFound.dataset
+        vesselMatchId = vesselFound.vesselMatchId
+      } else {
+        if (vessel.type === 'carrier') {
+          dataset = 'public-global-carrier-vessels:v20201001';
+        } else if (vessel.type === 'support') {
+          dataset = 'public-global-support-vessels:v20201001';
+        } else if (vessel.type === 'other') {
+          dataset = 'private-global-other-vessels:v20201001';
+        }
+      }
+      setProfileLoading(false)
+      dispatch(
+        redirect({
+          type: PROFILE,
+          payload: {
+            dataset: dataset,
+            vesselID: vessel.id ?? 'NA',
+            tmtID: vesselMatchId ?? 'NA',
+          }
+        })
+      )
+    },
+    [dispatch, findVessel]
+  )
+  
+  const onEncounterClick = useCallback(() => {
+    openVesselProfile(event.encounter?.vessel)
+  }, [openVesselProfile, event])
 
   return (
     <Fragment>
@@ -24,7 +69,10 @@ const ActivityModalContentDetailsEncounter: React.FC<ActivityModalContentProps> 
         <div className={styles.row}>
           <ActivityModalContentField
             label={t('vessel.encounteredVessel', 'Encountered Vessel')}
-            value={relatedVessel.name}
+            value={<span>
+              {relatedVessel.name} {profileLoading && <Spinner size="tiny" className={styles.profileLoader}></Spinner>}
+            </span>}
+            onValueClick={() => onEncounterClick()}
           />
           <ActivityModalContentField label={t('vessel.flag', 'Flag')} value={relatedVessel.flag} />
           <ActivityModalContentField label={t('vessel.mmsi', 'Mmsi')} value={relatedVessel.ssvid} />