+++ packages/api-client/src/utils/search.ts
@@ -13,6 +13,8 @@ export type AdvancedSearchQueryFieldKey =
   | 'imo'
   | 'callsign'
   | 'flag'
+  | 'geartype'
+  | 'target_species'
   | 'fleet'
   | 'origin'
   | 'lastTransmissionDate'
@@ -45,6 +47,14 @@ const FIELDS_PARAMS: Record<AdvancedSearchQueryFieldKey, AdvancedSearchQueryFiel
   },
   callsign: {
     operator: '=',
+  },
+  geartype: {
+    operator: 'IN',
+    transformation: multiSelectOptionToQuery,
+  },
+  target_species: {
+    operator: 'IN',
+    transformation: multiSelectOptionToQuery,
   },
   flag: {
     operator: 'IN',