+++ packages/timebar/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @globalfishingwatch/timebar
 
+## 2.3.0
+
+### Minor Changes
+
+- 8978b00d: range-selector-validity-checks
+
+## 2.2.0
+
+### Minor Changes
+
+- 0dd8b81d: replaced timebar date pickers with ui-component ones
+
 ## 2.1.0
 
 ### Minor Changes